Create visualizer for maze
Given an input array of matrices (each element of matrix is either 'W' for wall, 'T' for traversed, 'N' for neighbor, or a float denoting the cost of exploring the cell), write a function to make a gif of the maze traversal (see document for examples).
Document the required dependencies and how to install.